Add security csrf for delete pin

+++ src/Controller/PinsController.php
@@ -109,10 +109,14 @@ class PinsController extends AbstractController
     /**
      * @Route("/pins/{id<[0-9]+>}/delete", name="app_pins_delete", methods={"DELETE"})
      */
-    public function delete(Pin $pin, EntityManagerInterface $em): Response
+    public function delete(Request $request, Pin $pin, EntityManagerInterface $em): Response
     {
-        $em->remove($pin);
-        $em->flush();
+        $csrf = $request->request->get('csrf_token');
+        if ($this->isCsrfTokenValid('pin_deletion_' . $pin->getId(), $csrf)) {
+            
+            $em->remove($pin);
+            $em->flush();
+        }
 
         return $this->redirectToRoute('app_home');
     }
